feat: token 관리 파일 분리
